requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lark-oapi>=1.2.0
schedule>=1.2.0
icalendar>=5.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0

# creative bot
pyyaml>=6.0

# sentiment bot
jieba>=0.42.1
apscheduler>=3.10.0
PyGithub>=2.1.0

# newsbot
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
